third_party/blink/web_tests/external/wpt/css/css-anchor-position/anchor-getComputedStyle-002-expected.txt

This is a testharness.js-based test.
[FAIL] getComputedStyle() with fragmented containing block in multicolumn layout
  assert_equals: expected "25px" but got "-85px"
Harness: the test ran to completion.
